move test tools into Sofa.Testing directly and create Sofa.Testing.Tools target
