+++ Component/ODESolver/CMakeLists.txt
@@ -32,8 +32,6 @@ sofa_create_package_with_targets(
 
 cmake_dependent_option(SOFANG_ENABLE_TESTING_SOFA_COMPONENT_ODESOLVER "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
 if(SOFANG_ENABLE_TESTING_SOFA_COMPONENT_ODESOLVER)
-    # add testing framework
-    add_subdirectory(Testing)
     # add tests
     add_subdirectory(tests)
 endif()
