+++ SofaKernel/modules/Sofa.Testing/CMakeLists.txt
@@ -60,6 +60,8 @@ sofa_create_package_with_targets(
 
 set_target_properties(${PROJECT_NAME} PROPERTIES FOLDER Testing)
 
-if(SOFA_BUILD_TESTS)
-    add_subdirectory(${PROJECT_NAME}_test)
-endif()
+# test the tests...
+add_subdirectory(${PROJECT_NAME}_test)
+
+# add header-only tools (test templates, etc)
+add_subdirectory(tools)
