+++ SofaKernel/modules/Sofa.Testing/tools/src/sofa/testing/tools/empty.cpp
